// src/types.ts
export interface FormData {
    firstName?: string;
    lastName?: string;
    dob?: string;
    email?: string;
    phone?: string;
    loanAmount?: number;
    upfrontPayment?: number;
    terms?: number;
    monthlySalary?: number;
    additionalIncome?: number;
    mortgage?: number;
    otherCredits?: number;
  }